Refactor pickup exception retry flow to shared service
